skills/proposal-generator-skill/pyproject.toml
[project]
name = "proposal-generator"
version = "0.1.0"
description = "Generate Snowflake PS proposal presentations from meeting notes"
requires-python = ">=3.11"
dependencies = [
    "python-pptx>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[tool.uv]
dev-dependencies = []
